requirements.txt
Django==3.0
gunicorn==20.0.4
django-heroku==0.3.1
djangorestframework==3.11.0
django-admin-sortable==2.2.3
